backend/requirements.txt
Django==5.1
djangorestframework==3.15
python-dotenv==1.0
openai==1.50
django-cors-headers==4.4
gunicorn==23.0